Add support selecting a timezone and using it in the plot labels.
A project and a dataset can have multiple time zones. If the
dataset does not have a list of time zones, the project time zones
are used.

These timezones are provided in a selector widget when the user
is choosing a time.

Plotting javascript was updated to be timezone aware. Requires
latest version of Highcharts, or at least something more new than
highstock v2.0.4.  v2.1.5 works.