May have finally found solution to bug related to problems accessing chart.series[].data
Ajax function for datetime line StockChart usually found that
chart.series[].data.length was zero, even though the chart held data.
The bug disappears when the plot is created with highcharts({}) instead of
highcharts('StockChart',{}).  Plot appearance and functionality seems the same,
not sure that the "StockChart" added.

As part of this, use longer names for document time, data variables set in
dataset template.