Refactor to support single quotes in variable names
ISFS has NetCDF short_name attributes for variables, specifying the
prefered variable name, not subject to the NetCDF restrictions.
These typically have single quotes to indicate deviations:
    "w'h2o'.30m" is a covariance of w and h2o
Passing these to javascript in a django_template is a challenge:
    var plot_data = jQuery.parseJSON('{{ data }}');
The single quotes within data cause grief. The only way I know
around it is to replace the single quotes with \u0027.
I want to avoid that replacement on large data objects.
Therefore the data chunks are now lists instead of dicts, and the
mapping from variable name to list index in another dictionary,
which since it is small, replacing the quotes has less of an impact.

Also reworked the tab logic in dataset.html a bit.