Simplified the generation of tabs for the variable checkboxes
The elements in the tab dictionary are references to the
django.forms.widgets.CheckboxChoiceInput in the original
form['variables'], so they can be rendered in the template
without doing a lookup by variable name.

The make_tabs function is invoked from a template filter,
so it doesn't have to be called in the view.

Next we can add the create of tabs containing the variable category,
such as "met", "rad", etc when the platform for a dataset is ISFS.
Maybe make it a method in class Platform?