Copy attrs from NetCDF variables to model field
The variables field of Dataset has been specified for the
weather stations in json form, since the NetCDF files don't
have attributes such as units or long_names, and there are
some variables I don't think we want to provide.  But we need
to copy other attributes from the NetCDF variables, such as
dimnames and shape.