Fix bug in stnnames if variable not found in file
If a variable was not found in the last file read, the
value returned by stnnames was empty, which would cause
the variable not be plotted, and, I believe, but am not
completely sure, for the entire plot to be empty.

Add some content to log messages for bad times.