Simplify configuration a bit, update READMEs
Don't repeat the DATABASES configuration in settings/production.py,
instead just update the NAME if sqlite is used.

Remove __init__.py from datavis/settings, so that it isn't treated as a
package. This should make things a bit more clear.  DJANGO_SETTINGS_MODULE
should be set to default.py or production.py, not to the settings package.