Updates for django 2
Given light testing to see that it works in django 2.2.6, and is
backward compatible with django 1.11.17.

The only problem remaining with django 2.2.6 was with the
django-datetime-widget, where a "renderer=None" argument must be added
to the render() method in datetimewidget/widgets.py.
This fix was found in a pull request at
https://github.com/asaglimbeni/django-datetime-widget/, dated Feb 3 2019,
but not yet merged. widgets.py file is installed to lib/python3.x/site-packages
in the virtual environment, if one wants to hand edit it.